% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBTL.R
\name{dBTL}
\alias{dBTL}
\title{Calculate the density of ordinal comparison data based on BTL models}
\usage{
dBTL(Pi, omega, log = FALSE, groupwise = FALSE)
}
\arguments{
\item{Pi}{A matrix of rankings, such that the (i,j) entry is the index of the jth-most preferred object according to judge i. If \code{groupwise=TRUE}, then the index corresponds to the jth-most preferred object among those in row i; if \code{groupwise=FALSE}, it is assumd that all unranked objects (if any) are less preferred than those which are ranked.}

\item{omega}{A vector of non-negative object worth parameters.}

\item{log}{A boolean to indicate whether the log density should be returned. Default to \code{FALSE}.}

\item{groupwise}{A boolean to indicate whether the observed rankings are complete/partial rankings (\code{FALSE}; default) or groupwise comparisons (\code{TRUE}).}
}
\value{
A numeric joint density of the observed rankings.
}
\description{
This function calculates the density of ordinal comparison data (e.g., complete rankings, partial rankings, pairwise comparisons, or groupwise comparisons) based on an appropriate Bradley-Terry-Luce (BTL) model with object-specific worth parameters, \code{omega}.
}
\examples{
Pi <- matrix(data=c(1,2,3,NA,NA,1,2,3,4,5),byrow=TRUE,nrow=2)
dBTL(Pi=Pi,omega=c(10,8,6,4,2),log=TRUE,groupwise=FALSE)
dBTL(Pi=Pi,omega=c(10,8,6,4,2),log=TRUE,groupwise=TRUE)
}
