% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_BTL.R
\name{fit_BTL}
\alias{fit_BTL}
\title{Fit a Bayesian BTL model to ordinal comparison data (recommended for internal use only; use mcmc_BTL function instead)}
\usage{
fit_BTL(
  Pi,
  J,
  a_gamma,
  b_gamma,
  nu0 = NULL,
  num_iters = 100,
  groupwise = FALSE
)
}
\arguments{
\item{Pi}{A matrix of rankings, such that the (i,j) entry is the index of the jth-most preferred object according to judge i. If \code{groupwise=TRUE}, then the index corresponds to the jth-most preferred object among those in row i; if \code{groupwise=FALSE}, it is assumd that all unranked objects (if any) are less preferred than those which are ranked.}

\item{J}{A numeric indicating the total number of objects being compared.}

\item{a_gamma}{A numeric for the first hyperparameter (shape) in a Gamma prior on each worth parameter.}

\item{b_gamma}{A numeric for the second hyperparameter (rate) in a Gamma prior on each worth parameter.}

\item{nu0}{A numeric vector for the initialization of worth parameters, omega, in the MCMC algorithm. Default to \code{NULL}, indicating random initialization.}

\item{num_iters}{A numeric indicating the total number of MCMC iterations.}

\item{groupwise}{A boolean to indicate whether the observed rankings are complete/partial rankings (\code{FALSE}; default) or groupwise comparisons (\code{TRUE}).}
}
\value{
A list containing a single element, \code{omega}, a \code{num_iters}x\code{J} matrix of approximate posterior draws of the object-specific worth parameters, omega.
}
\description{
This function fits a standard Bayesian BTL model to ordinal comparison data (e.g., complete rankings, partial rankings, pairwise comparisons, or groupwise comparisons) such that each worth parameter receives an independent Gamma prior.
}
\examples{
Pi <- matrix(data=c(1,2,3,NA,NA,1,2,3,4,5),byrow=TRUE,nrow=2)
fit_BTL(Pi=Pi,J=5,a_gamma=5,b_gamma=3,num_iters=10)
}
