% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_RCBTL.R
\name{mcmc_RCBTL}
\alias{mcmc_RCBTL}
\title{Fit Bayesian Rank-Clustered BTL model to ordinal comparison data using multiple MCMC chains}
\usage{
mcmc_RCBTL(
  Pi,
  J,
  a_gamma,
  b_gamma,
  lambda,
  nu0 = NULL,
  num_iters = 1000,
  nu_reps = 2,
  groupwise = FALSE,
  chains = 4,
  burn_prop = 0.5,
  thin = 1,
  seed = NULL,
  normalize_omega = TRUE
)
}
\arguments{
\item{Pi}{A matrix of preference orderings ("rankings"), such that the (i,j) entry is the index of the jth-most preferred object according to judge i. If \code{groupwise=TRUE}, then the index corresponds to the jth-most preferred object among those in row i; if \code{groupwise=FALSE}, it is assumed that all unranked objects (if any) are less preferred than those which are ranked.}

\item{J}{A numeric indicating the total number of objects being compared.}

\item{a_gamma}{A numeric for the first hyperparameter (shape) in a Gamma prior on each worth parameter.}

\item{b_gamma}{A numeric for the second hyperparameter (rate) in a Gamma prior on each worth parameter.}

\item{lambda}{A numeric for the Poisson hyperparameter on the number of non-empty clusters in the partition of worth parameters.}

\item{nu0}{A numeric vector for the initialization of worth parameters, omega, in the MCMC algorithm. Default to \code{NULL}, indicating random initialization.}

\item{num_iters}{A numeric indicating the total number of outer MCMC iterations (i.e., the number of times the partition is updated in the Gibbs sampler).}

\item{nu_reps}{A numeric indicating the number of times each worth parameter is drawn per update of the parameter partition. There will be a total of \code{num_iters}x\code{nu_reps} samples from the posterior.}

\item{groupwise}{A boolean to indicate whether the observed rankings are complete/partial rankings (\code{FALSE}; default) or groupwise comparisons (\code{TRUE}).}

\item{chains}{A numeric indicating the total number of independent MCMC chains to be run.}

\item{burn_prop}{A numeric between 0 and 1 indicating the proportion of MCMC samples in each chain to be removed as burn-in.}

\item{thin}{A numeric indicating that only every \code{thin}-th sample should be retained, to save computational memory.}

\item{seed}{A numeric indicating the random seed that should be set before running the first MCMC chain.}

\item{normalize_omega}{A boolean indicating if each posterior draw of omega should be normalized post-hoc to sum to 1; removes a standard identifiability concern of BTL models.}
}
\value{
A (\code{chains}x\code{num_iters}/\code{thin})x(3J+3) matrix of posterior draws, one row per posterior sample of omega, nu, and g, with additional columns indicating the MCMC chain index, iteration index, and number of non-empty partition clusters K of each posterior sample.
}
\description{
This function fits a Bayesian Rank-Clustered BTL model to ordinal comparison data (e.g., complete rankings, partial rankings, pairwise comparisons, or groupwise comparisons) such that the worth parameters are drawn from a PSSF prior (as defined in Pearce and Erosheva 2024). The function has input parameters to permit drawing multiple MCMC chains, as well as thinning and burn-in.
}
\examples{
Pi <- matrix(data=c(1,2,3,NA,NA,1,2,3,4,5),byrow=TRUE,nrow=2)
mcmc_RCBTL(Pi=Pi,J=5,a_gamma=5,b_gamma=3,lambda=2,num_iters=6,chains=2,seed=1)
}
