% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_constants.R
\name{obtain_constants}
\alias{obtain_constants}
\title{Calculation of constants for Gibbs sampling of worth parameters via data augmentation (internal use only)}
\usage{
obtain_constants(Pi, I, J, groupwise = FALSE)
}
\arguments{
\item{Pi}{A matrix of preference orderings ("rankings"), such that the (i,j) entry is the index of the jth-most preferred object according to judge i. If \code{groupwise=TRUE}, then the index corresponds to the jth-most preferred object among those in row i; if \code{groupwise=FALSE}, it is assumed that all unranked objects (if any) are less preferred than those which are ranked.}

\item{I}{A numeric indicating the number of rows in Pi}

\item{J}{A numeric indicating the total number of objects being compared.}

\item{groupwise}{A boolean to indicate whether the observed rankings are complete/partial rankings (\code{FALSE}; default) or groupwise comparisons (\code{TRUE}).}
}
\value{
A list containing calculated constants c_k0 and delta_irk0.
}
\description{
This function calculates constants c_k0 and delta_irk0 based on the observed rankings \code{Pi}, as defined in Pearce and Erosheva (2024). For internal use only.
}
