% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_density.R
\name{posterior_density}
\alias{posterior_density}
\title{Calculate proportional posterior density for mcmc samples in a Rank-Clustered BTL model.}
\usage{
posterior_density(
  Pi,
  mcmc,
  a_gamma,
  b_gamma,
  lambda,
  groupwise = FALSE,
  log = FALSE
)
}
\arguments{
\item{Pi}{A matrix of preference orderings ("rankings"), such that the (i,j) entry is the index of the jth-most preferred object according to judge i. If \code{groupwise=TRUE}, then the index corresponds to the jth-most preferred object among those in row i; if \code{groupwise=FALSE}, it is assumed that all unranked objects (if any) are less preferred than those which are ranked.}

\item{mcmc}{The matrix output of the \code{mcmc_BTL} or \code{mcmc_RCBTL} functions.}

\item{a_gamma}{A numeric for the first hyperparameter (shape) in a Gamma prior on each worth parameter.}

\item{b_gamma}{A numeric for the second hyperparameter (rate) in a Gamma prior on each worth parameter.}

\item{lambda}{A numeric for the Poisson hyperparameter on the number of non-empty clusters in the partition of worth parameters.}

\item{groupwise}{A boolean to indicate whether the observed rankings are complete/partial rankings (\code{FALSE}; default) or groupwise comparisons (\code{TRUE}).}

\item{log}{A boolean to indicate whether the log posterior densities should be returned. Default to \code{FALSE}.}
}
\value{
A ggplot(s) of side-by-side violin plots for each omega_j.
}
\description{
This function calculates the (proportional) posterior density for MCMC samples from a Rank-Clustered BTL model.
}
\examples{
Pi <- rBTL(I=30,omega=c(4^2,4^2,4^1,4^1,4^0))
mcmc <- mcmc_RCBTL(Pi=Pi,J=5,a_gamma=5,b_gamma=3,lambda=1,num_iters=200,chains=2,seed=1)
posterior_density(Pi = Pi, mcmc = mcmc, a_gamma=5, b_gamma = 3, lambda=2, log=TRUE)
}
