% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_ggplots.R
\name{posterior_ggplots}
\alias{posterior_ggplots}
\title{Posterior ggplots for worth parameters in (Rank-Clustered) BTL models.}
\usage{
posterior_ggplots(mcmc, object_names = NULL)
}
\arguments{
\item{mcmc}{The matrix output of the \code{mcmc_BTL} or \code{mcmc_RCBTL} functions.}

\item{object_names}{An optional vector of object names, in the same index order as the observed data.}
}
\value{
A ggplot(s) of side-by-side violin plots for each omega_j.
}
\description{
This function creates side-by-side violin plots for omega (object-specific worth parameters) using MCMC chains from a BTL or Rank-Clustered BTL distribution.
}
\examples{
Pi <- rBTL(I=30,omega=c(4^2,4^2,4^1,4^1,4^0))
mcmc <- mcmc_BTL(Pi=Pi,J=5,a_gamma=5,b_gamma=3,num_iters=200,chains=2,seed=1)
posterior_ggplots(mcmc,object_names=paste0("Object ",1:5))
}
