% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBTL.R
\name{rBTL}
\alias{rBTL}
\title{Random data generation from standard BTL models}
\usage{
rBTL(I, omega, R = NULL)
}
\arguments{
\item{I}{A numeric indicating the number of observations to draw.}

\item{omega}{A vector of non-negative object worth parameters.}

\item{R}{A numeric indicating the length of each ranking. Default to \code{NULL}, indicating to draw complete rankings.}
}
\value{
A \code{I}x\code{R} matrix of rankings.
}
\description{
This function randomly draws data from a Bradley-Terry-Luce (BTL) model with object-specific worth parameters, \code{omega}. Currently, only complete or partial rankings may be drawn, although those may be subsetted to simulate groupwise comparison data as a consequence of Luce's Axiom of Choice.
}
\examples{
set.seed(1)
rBTL(I=5,omega=c(10,5,1),R=3)
}
