% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRCBTL.R
\name{rRCBTL}
\alias{rRCBTL}
\title{Random data generation from Rank-Clustered BTL models}
\usage{
rRCBTL(I, J, a_gamma, b_gamma, lambda, return_all = FALSE)
}
\arguments{
\item{I}{A numeric indicating the number of observations to draw.}

\item{J}{A numeric indicating the total number of objects being compared.}

\item{a_gamma}{A numeric for the first hyperparameter (shape) in a Gamma prior on each worth parameter.}

\item{b_gamma}{A numeric for the second hyperparameter (rate) in a Gamma prior on each worth parameter.}

\item{lambda}{A numeric for the Poisson hyperparameter on the number of non-empty clusters in the partition of worth parameters.}

\item{return_all}{A boolean indicator if K, partitions, nu, and omega should be return in addition to the generated rankings, Pi.}
}
\value{
A \code{I}x\code{R} matrix of rankings (if \code{return_all=FALSE}), or a list containing Pi, K, partitions, nu, and omega (if \code{return_all=TRUE}).
}
\description{
This function randomly draws data from a Rank-Clustered Bradley-Terry-Luce (BTL) model with object-specific worth parameters, \code{omega}. Currently, only complete or partial rankings may be drawn, although those may be subsetted to simulate groupwise comparison data as a consequence of Luce's Axiom of Choice.
}
\examples{
set.seed(1)
rRCBTL(I=3,J=5,a_gamma=5,b_gamma=3,lambda=3,return_all=TRUE)
}
