% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_nu.R
\name{sample_nu}
\alias{sample_nu}
\title{Sampling of unique parameter worth values, nu, in Bayesian Rank-Clustered BTL models (internal use only)}
\usage{
sample_nu(
  n = 1,
  Pi,
  I,
  J,
  nu,
  p,
  K,
  a_gamma,
  b_gamma,
  c_k0 = NULL,
  delta_irk0 = NULL,
  groupwise = FALSE
)
}
\arguments{
\item{n}{A numeric indicating the number of posterior samples to draw.}

\item{Pi}{A matrix of rankings, such that the (i,j) entry is the index of the jth-most preferred object according to judge i. If \code{groupwise=TRUE}, then the index corresponds to the jth-most preferred object among those in row i; if \code{groupwise=FALSE}, it is assumd that all unranked objects (if any) are less preferred than those which are ranked.}

\item{I}{A numeric indicating the number of rows in Pi}

\item{J}{A numeric indicating the total number of objects being compared.}

\item{nu}{A vector indicating current values for nu in the Gibbs sampler.}

\item{p}{A vector indicating current values for g in the Gibbs sampler.}

\item{K}{A vector indicating current values for K in the Gibbs sampler.}

\item{a_gamma}{A numeric for the first hyperparameter (shape) in a Gamma prior on each worth parameter.}

\item{b_gamma}{A numeric for the second hyperparameter (rate) in a Gamma prior on each worth parameter.}

\item{c_k0}{A vector of constants, as calculated by the obtain_constants function.}

\item{delta_irk0}{A matrix of constants, as calculated by the obtain_constants function.}

\item{groupwise}{A boolean to indicate whether the observed rankings are complete/partial rankings (\code{FALSE}; default) or groupwise comparisons (\code{TRUE}).}
}
\value{
A matrix containing updated values for nu for n posterior samples.
}
\description{
This function implements a Gibbs sampler via data augmentation for updating the unique worth parameters, nu, in Bayesian Rank-Clustered BTL models. For internal use only.
}
